[GameRunner2 Extract Methods] Move getRootFolder method from GameRunner2 to ClientContext and fix ciruclar method call. VERY IMPORTANT - fix the call to 'getTripleaJarWithEngineVersionStringPath' to not need game engine version. That introduced a circular depdency. To read the game engine configuration file, we need to know the game engine version, to know the game engine version, we need to read the game engine configuration file.
